+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ==================
 
 * FindByLabelText
+* LocalStorage.all
 * Element getAttribute method
 
 2.9.0 / 2019-11-20