+++ CHANGELOG.md
@@ -51,7 +51,7 @@
 2.6.0 / 2019-08-06
 ==================
 
-* Browser.text will return newlines on </br>
+* Browser.text will return newlines on `</br>`
 * DefaultTimeout option on create browser
 * Injection scripts are now added through evaluate, so are not checked by CSP
 * Static config file removed